✨ 서버에 이모지를 등록할 때 이모지 현황이 정상적으로 갱신되도록 함
