+++ src/pages/post/components/HeaderButtonBox/HeaderButtonBox.jsx
@@ -1,13 +1,13 @@
 import HeaderAddEmojiButton from "../HeaderAddEmojiButton/HeaderAddEmojiButton";
 import HeaderShareButton from "../HeaderShareButton/HeaderShareButton";
 import * as S from "./HeaderButtonBox.style";
-const HeaderButtonBox = () => {
-  return (
-    <S.Box>
-      <HeaderAddEmojiButton />
-      <HeaderShareButton />
-    </S.Box>
-  );
+const HeaderButtonBox = ({ emojiFunc }) => {
+	return (
+		<S.Box>
+			<HeaderAddEmojiButton emojiFunc={emojiFunc} />
+			<HeaderShareButton />
+		</S.Box>
+	);
 };
 
 export default HeaderButtonBox;
